requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.10
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
Django==5.1.6
docopt==0.6.2
elevenlabs==1.53.0
filelock==3.15.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.48.1
fsspec==2024.6.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
keyboard==0.13.5
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.20
mkl==2021.4.0
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.0.8
narwhals==1.23.0
networkx==3.3
nltk==3.8.1
numpy==1.26.3
openai==1.65.4
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
packaging==23.2
pandas==2.2.0
pandass==1.11.4
pillow==10.2.0
pipwin==0.5.2
protobuf==4.25.5
pyarrow==15.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyparsing==3.1.1
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python_speech_features==0.6
pyttsx3==2.98
pytweening==1.2.0
pytz==2023.4
pywhatkit==5.4
pywin32==308
PyYAML==6.0.1
referencing==0.36.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.22.3
safetensors==0.4.3
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.6.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
SpeechRecognition==3.14.1
sqlparse==0.5.3
streamlit==1.41.1
sympy==1.12.1
tbb==2021.12.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tornado==6.4.2
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.3.1
urllib3==2.2.1
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2